+++ packages/vscode-extension/src/contextService.ts
@@ -1,5 +1,5 @@
 import { Neo4jConnectionSettings } from '@neo4j-cypher/language-server/src/types';
-import { Neo4jSchemaPoller } from '@neo4j-cypher/schema-poller';
+import { Neo4jSchemaPoller } from '@neo4j-cypher/query-tools';
 import { ExtensionContext } from 'vscode';
 import CypherRunner from './cypherRunner';